refactor(contracts): move witnesses into per-module test/witnesses/ (#528)
