+++ contracts/src/access/test/witnesses/AccessControlWitnesses.ts
@@ -1,39 +1,39 @@
 // SPDX-License-Identifier: MIT
-// OpenZeppelin Compact Contracts v0.0.1-alpha.1 (token/witnesses/NonFungibleToken.ts)
+// OpenZeppelin Compact Contracts v0.0.1-alpha.1 (access/test/witnesses/AccessControlWitnesses.ts)
 
 import { getRandomValues } from 'node:crypto';
 import type { WitnessContext } from '@midnight-ntwrk/compact-runtime';
 
 /**
- * @description Interface defining the witness methods for NonFungibleToken operations.
+ * @description Interface defining the witness methods for AccessControl operations.
  * @template P - The private state type.
  */
-export interface INonFungibleTokenWitnesses<L, P> {
+export interface IAccessControlWitnesses<L, P> {
   /**
    * Retrieves the secret key from the private state.
    * @param context - The witness context containing the private state.
    * @returns A tuple of the private state and the secret key as a Uint8Array.
    */
-  wit_NonFungibleTokenSK(context: WitnessContext<L, P>): [P, Uint8Array];
+  wit_AccessControlSK(context: WitnessContext<L, P>): [P, Uint8Array];
 }
 
 /**
- * @description Represents the private state of an NonFungibleToken contract, storing a secret key.
+ * @description Represents the private state of an AccessControl contract, storing a secret key.
  */
-export type NonFungibleTokenPrivateState = {
+export type AccessControlPrivateState = {
   /** @description A 32-byte secret key used for creating a public user identifier. */
   secretKey: Uint8Array;
 };
 
 /**
- * @description Utility object for managing the private state of an NonFungibleToken contract.
+ * @description Utility object for managing the private state of an AccessControl contract.
  */
-export const NonFungibleTokenPrivateState = {
+export const AccessControlPrivateState = {
   /**
    * @description Generates a new private state with a random secret key.
-   * @returns A fresh NonFungibleTokenPrivateState instance.
+   * @returns A fresh AccessControlPrivateState instance.
    */
-  generate: (): NonFungibleTokenPrivateState => {
+  generate: (): AccessControlPrivateState => {
     return { secretKey: getRandomValues(new Uint8Array(32)) };
   },
 
@@ -42,16 +42,16 @@ export const NonFungibleTokenPrivateState = {
    * Useful for deterministic key generation or advanced use cases.
    *
    * @param sk - The 32-byte secret key to use.
-   * @returns A fresh NonFungibleTokenPrivateState instance with the provided key.
+   * @returns A fresh AccessControlPrivateState instance with the provided key.
    *
    * @example
    * ```typescript
    * // For deterministic keys (user-defined scheme)
    * const deterministicKey = myDeterministicScheme(...);
-   * const privateState = NonFungibleTokenPrivateState.withSecretKey(deterministicKey);
+   * const privateState = AccessControlPrivateState.withSecretKey(deterministicKey);
    * ```
    */
-  withSecretKey: (sk: Uint8Array): NonFungibleTokenPrivateState => {
+  withSecretKey: (sk: Uint8Array): AccessControlPrivateState => {
     if (sk.length !== 32) {
       throw new Error(
         `withSecretKey: expected 32-byte secret key, received ${sk.length} bytes`,
@@ -62,16 +62,16 @@ export const NonFungibleTokenPrivateState = {
 };
 
 /**
- * @description Factory function creating witness implementations for NonFungibleToken operations.
- * @returns An object implementing the Witnesses interface for NonFungibleTokenPrivateState.
+ * @description Factory function creating witness implementations for AccessControl operations.
+ * @returns An object implementing the Witnesses interface for AccessControlPrivateState.
  */
-export const NonFungibleTokenWitnesses = <L>(): INonFungibleTokenWitnesses<
+export const AccessControlWitnesses = <L>(): IAccessControlWitnesses<
   L,
-  NonFungibleTokenPrivateState
+  AccessControlPrivateState
 > => ({
-  wit_NonFungibleTokenSK(
-    context: WitnessContext<L, NonFungibleTokenPrivateState>,
-  ): [NonFungibleTokenPrivateState, Uint8Array] {
+  wit_AccessControlSK(
+    context: WitnessContext<L, AccessControlPrivateState>,
+  ): [AccessControlPrivateState, Uint8Array] {
     return [
       context.privateState,
       Uint8Array.from(context.privateState.secretKey),
