+++ contracts/src/archive/test/witnesses/ShieldedTokenWitnesses.ts
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-// OpenZeppelin Compact Contracts v0.0.1-alpha.1 (archive/witnesses/ShieldedTokenWitnesses.ts)
+// OpenZeppelin Compact Contracts v0.0.1-alpha.1 (archive/test/witnesses/ShieldedTokenWitnesses.ts)
 
 // This is how we type an empty object.
 export type ShieldedTokenPrivateState = Record<string, never>;
