+++ contracts/src/security/test/witnesses/InitializableWitnesses.ts
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-// OpenZeppelin Compact Contracts v0.0.1-alpha.1 (security/witnesses/InitializableWitnesses.ts)
+// OpenZeppelin Compact Contracts v0.0.1-alpha.1 (security/test/witnesses/InitializableWitnesses.ts)
 
 export type InitializablePrivateState = Record<string, never>;
 export const InitializablePrivateState: InitializablePrivateState = {};
