+++ contracts/src/security/test/witnesses/PausableWitnesses.ts
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-// OpenZeppelin Compact Contracts v0.0.1-alpha.1 (security/witnesses/PausableWitnesses.ts)
+// OpenZeppelin Compact Contracts v0.0.1-alpha.1 (security/test/witnesses/PausableWitnesses.ts)
 
 export type PausablePrivateState = Record<string, never>;
 export const PausablePrivateState: PausablePrivateState = {};
