+++ contracts/src/token/test/witnesses/MultiTokenWitnesses.ts
@@ -1,39 +1,39 @@
 // SPDX-License-Identifier: MIT
-// OpenZeppelin Compact Contracts v0.0.1-alpha.1 (access/witnesses/AccessControlWitnesses.ts)
+// OpenZeppelin Compact Contracts v0.0.1-alpha.1 (token/test/witnesses/MultiTokenWitnesses.ts)
 
 import { getRandomValues } from 'node:crypto';
 import type { WitnessContext } from '@midnight-ntwrk/compact-runtime';
 
 /**
- * @description Interface defining the witness methods for AccessControl operations.
+ * @description Interface defining the witness methods for MultiToken operations.
  * @template P - The private state type.
  */
-export interface IAccessControlWitnesses<L, P> {
+export interface IMultiTokenWitnesses<L, P> {
   /**
    * Retrieves the secret key from the private state.
    * @param context - The witness context containing the private state.
    * @returns A tuple of the private state and the secret key as a Uint8Array.
    */
-  wit_AccessControlSK(context: WitnessContext<L, P>): [P, Uint8Array];
+  wit_MultiTokenSK(context: WitnessContext<L, P>): [P, Uint8Array];
 }
 
 /**
- * @description Represents the private state of an AccessControl contract, storing a secret key.
+ * @description Represents the private state of an MultiToken contract, storing a secret key.
  */
-export type AccessControlPrivateState = {
+export type MultiTokenPrivateState = {
   /** @description A 32-byte secret key used for creating a public user identifier. */
   secretKey: Uint8Array;
 };
 
 /**
- * @description Utility object for managing the private state of an AccessControl contract.
+ * @description Utility object for managing the private state of an MultiToken contract.
  */
-export const AccessControlPrivateState = {
+export const MultiTokenPrivateState = {
   /**
    * @description Generates a new private state with a random secret key.
-   * @returns A fresh AccessControlPrivateState instance.
+   * @returns A fresh MultiTokenPrivateState instance.
    */
-  generate: (): AccessControlPrivateState => {
+  generate: (): MultiTokenPrivateState => {
     return { secretKey: getRandomValues(new Uint8Array(32)) };
   },
 
@@ -42,16 +42,16 @@ export const AccessControlPrivateState = {
    * Useful for deterministic key generation or advanced use cases.
    *
    * @param sk - The 32-byte secret key to use.
-   * @returns A fresh AccessControlPrivateState instance with the provided key.
+   * @returns A fresh MultiTokenPrivateState instance with the provided key.
    *
    * @example
    * ```typescript
    * // For deterministic keys (user-defined scheme)
    * const deterministicKey = myDeterministicScheme(...);
-   * const privateState = AccessControlPrivateState.withSecretKey(deterministicKey);
+   * const privateState = MultiTokenPrivateState.withSecretKey(deterministicKey);
    * ```
    */
-  withSecretKey: (sk: Uint8Array): AccessControlPrivateState => {
+  withSecretKey: (sk: Uint8Array): MultiTokenPrivateState => {
     if (sk.length !== 32) {
       throw new Error(
         `withSecretKey: expected 32-byte secret key, received ${sk.length} bytes`,
@@ -62,16 +62,16 @@ export const AccessControlPrivateState = {
 };
 
 /**
- * @description Factory function creating witness implementations for AccessControl operations.
- * @returns An object implementing the Witnesses interface for AccessControlPrivateState.
+ * @description Factory function creating witness implementations for MultiToken operations.
+ * @returns An object implementing the Witnesses interface for MultiTokenPrivateState.
  */
-export const AccessControlWitnesses = <L>(): IAccessControlWitnesses<
+export const MultiTokenWitnesses = <L>(): IMultiTokenWitnesses<
   L,
-  AccessControlPrivateState
+  MultiTokenPrivateState
 > => ({
-  wit_AccessControlSK(
-    context: WitnessContext<L, AccessControlPrivateState>,
-  ): [AccessControlPrivateState, Uint8Array] {
+  wit_MultiTokenSK(
+    context: WitnessContext<L, MultiTokenPrivateState>,
+  ): [MultiTokenPrivateState, Uint8Array] {
     return [
       context.privateState,
       Uint8Array.from(context.privateState.secretKey),
