+++ contracts/src/token/test/witnesses/NonFungibleTokenWitnesses.ts
@@ -1,39 +1,39 @@
 // SPDX-License-Identifier: MIT
-// OpenZeppelin Compact Contracts v0.0.1-alpha.1 (access/witnesses/OwnableWitnesses.ts)
+// OpenZeppelin Compact Contracts v0.0.1-alpha.1 (token/test/witnesses/NonFungibleTokenWitnesses.ts)
 
 import { getRandomValues } from 'node:crypto';
 import type { WitnessContext } from '@midnight-ntwrk/compact-runtime';
 
 /**
- * @description Interface defining the witness methods for Ownable operations.
+ * @description Interface defining the witness methods for NonFungibleToken operations.
  * @template P - The private state type.
  */
-export interface IOwnableWitnesses<L, P> {
+export interface INonFungibleTokenWitnesses<L, P> {
   /**
    * Retrieves the secret key from the private state.
    * @param context - The witness context containing the private state.
    * @returns A tuple of the private state and the secret key as a Uint8Array.
    */
-  wit_OwnableSK(context: WitnessContext<L, P>): [P, Uint8Array];
+  wit_NonFungibleTokenSK(context: WitnessContext<L, P>): [P, Uint8Array];
 }
 
 /**
- * @description Represents the private state of an Ownable contract, storing a secret key.
+ * @description Represents the private state of an NonFungibleToken contract, storing a secret key.
  */
-export type OwnablePrivateState = {
+export type NonFungibleTokenPrivateState = {
   /** @description A 32-byte secret key used for creating a public user identifier. */
   secretKey: Uint8Array;
 };
 
 /**
- * @description Utility object for managing the private state of an Ownable contract.
+ * @description Utility object for managing the private state of an NonFungibleToken contract.
  */
-export const OwnablePrivateState = {
+export const NonFungibleTokenPrivateState = {
   /**
    * @description Generates a new private state with a random secret key.
-   * @returns A fresh OwnablePrivateState instance.
+   * @returns A fresh NonFungibleTokenPrivateState instance.
    */
-  generate: (): OwnablePrivateState => {
+  generate: (): NonFungibleTokenPrivateState => {
     return { secretKey: getRandomValues(new Uint8Array(32)) };
   },
 
@@ -42,16 +42,16 @@ export const OwnablePrivateState = {
    * Useful for deterministic key generation or advanced use cases.
    *
    * @param sk - The 32-byte secret key to use.
-   * @returns A fresh OwnablePrivateState instance with the provided key.
+   * @returns A fresh NonFungibleTokenPrivateState instance with the provided key.
    *
    * @example
    * ```typescript
    * // For deterministic keys (user-defined scheme)
    * const deterministicKey = myDeterministicScheme(...);
-   * const privateState = OwnablePrivateState.withSecretKey(deterministicKey);
+   * const privateState = NonFungibleTokenPrivateState.withSecretKey(deterministicKey);
    * ```
    */
-  withSecretKey: (sk: Uint8Array): OwnablePrivateState => {
+  withSecretKey: (sk: Uint8Array): NonFungibleTokenPrivateState => {
     if (sk.length !== 32) {
       throw new Error(
         `withSecretKey: expected 32-byte secret key, received ${sk.length} bytes`,
@@ -62,16 +62,16 @@ export const OwnablePrivateState = {
 };
 
 /**
- * @description Factory function creating witness implementations for Ownable operations.
- * @returns An object implementing the Witnesses interface for OwnablePrivateState.
+ * @description Factory function creating witness implementations for NonFungibleToken operations.
+ * @returns An object implementing the Witnesses interface for NonFungibleTokenPrivateState.
  */
-export const OwnableWitnesses = <L>(): IOwnableWitnesses<
+export const NonFungibleTokenWitnesses = <L>(): INonFungibleTokenWitnesses<
   L,
-  OwnablePrivateState
+  NonFungibleTokenPrivateState
 > => ({
-  wit_OwnableSK(
-    context: WitnessContext<L, OwnablePrivateState>,
-  ): [OwnablePrivateState, Uint8Array] {
+  wit_NonFungibleTokenSK(
+    context: WitnessContext<L, NonFungibleTokenPrivateState>,
+  ): [NonFungibleTokenPrivateState, Uint8Array] {
     return [
       context.privateState,
       Uint8Array.from(context.privateState.secretKey),
