+++ contracts/src/utils/test/witnesses/UtilsWitnesses.ts
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-// OpenZeppelin Compact Contracts v0.0.1-alpha.1 (utils/witnesses/UtilsWitnesses.ts)
+// OpenZeppelin Compact Contracts v0.0.1-alpha.1 (utils/test/witnesses/UtilsWitnesses.ts)
 
 export type UtilsPrivateState = Record<string, never>;
 export const UtilsPrivateState: UtilsPrivateState = {};
